BugFix: In wxWidgets 2.6 there was a protection fault in "OnSelectedChanged"
git-svn-id: http://astade.tigris.org/svn/astade/trunk@885 aaa30fad-34f9-0310-8aa9-877cc762ee2e

+++ Model/package_SVSWQPEV/package_YGKSEMEQ/Package_ASFBWBFP/Package_DKNBOEDV/classes/class_RGNDUWKU/operations/operation_XTFYCSLB/code.cpp
@@ -1,4 +1,4 @@
-if (event.GetId()==pinID)
+if (static_cast<unsigned int>(event.GetId())==pinID)
 	m_Pinned = !m_Pinned;
 else
 	event.Skip();
